Add AuthenticationException to plugin api
- Move RealmLifeCycle to plugin spi
