+++ enterprise/security/src/main/java/org/neo4j/server/security/enterprise/auth/plugin/spi/AuthPlugin.java
@@ -21,7 +21,7 @@ package org.neo4j.server.security.enterprise.auth.plugin.spi;
 
 import java.util.Map;
 
-import org.neo4j.server.security.enterprise.auth.RealmLifecycle;
+import org.neo4j.server.security.enterprise.auth.plugin.api.AuthenticationException;
 
 /**
  * TODO
@@ -31,5 +31,5 @@ public interface AuthPlugin extends RealmLifecycle
     /**
      * TODO
      */
-    AuthInfo getAuthInfo( Map<String,Object> authToken );
+    AuthInfo getAuthInfo( Map<String,Object> authToken ) throws AuthenticationException;
 }
